Refactoring & fragment code cleaning
