+++ app/src/main/java/com/timermakov/homeworkmtsteta/fragments/MovieDetailsFragment.kt
@@ -1,0 +1,19 @@
+package com.timermakov.homeworkmtsteta.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.timermakov.homeworkmtsteta.R
+
+class MovieDetailsFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_movie_details, container, false)
+    }
+
+}
