+++ app/src/main/java/com/timermakov/homeworkmtsteta/movies/MoviesModel.kt
@@ -1,6 +1,4 @@
-package com.timermakov.homeworkmtsteta
-
-import com.timermakov.homeworkmtsteta.movies.MoviesDataSource
+package com.timermakov.homeworkmtsteta.movies
 
 class MoviesModel(private val moviesDataSource: MoviesDataSource) {
     fun getMovies() = moviesDataSource.getMovies()
